ml_final_project/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
tensorflow>=2.15.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
firebase-admin>=6.4.0
google-cloud-firestore>=2.13.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0